Separate MatPower from other test utilities.
