+++ examples/PowerFlow/MatPowerTesting/test_parse_branch_row.cpp
@@ -2,7 +2,7 @@
 
 #include <FileIO.hpp>
 #include <PowerSystemData.hpp>
-#include <Testing.hpp>
+#include "MatPowerTesting.hpp"
 
 using namespace GridKit;
 using namespace GridKit::Testing;
